requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-cleanup==5.0.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
wincertstore==0.2
zipp==3.1.0
